+++ CHANGELOG.md
@@ -20,9 +20,9 @@ Please mark all change in change log and use the issue from GitHub
 -   \#1547 Rename storage/file to storage/disk and rename classes
 -   \#1548 Move store/Directory to storage/Operation and add FSHandler
 -   \#1649 Fix Milvus crash on old CPU 
+-   \#1619 Improve compact performance
 
 ## Task
-
 
 # Milvus 0.7.0 (2020-03-11)